Style <select> nicer in linux
Mac chrome doesn't allow changing the dropdown style. In linux
inheriting the `color` attribute make the rendering of options pretty
terrible. Hardcode background color instead of inherit. Set background
color in :active state.

Move the arrow 'inside' the select box so clicking on it opens the
dropdown.

Add a -mox-appearance alongside -webkit-appearance in the hopes that
this works with firefox as well.